tests: Move CheckMessageTest to test_message_send.py.
